app1_sender/requirements.txt
pika==1.3.0 
pyyaml==5.4.1
Flask==2.1.3
